tests/wpt/tests/css/css-scroll-anchoring/parsing/overflow-anchor-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Scroll Anchoring: getComputedStyle().overflowAnchor</title>
<link rel="help" href="https://drafts.csswg.org/css-scroll-anchoring/#propdef-overflow-anchor">
<meta name="assert" content="overflow-anchor computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("overflow-anchor", "auto");
test_computed_value("overflow-anchor", "none");
</script>
</body>
</html>
